python/pyproject.toml
[project]
name = "fricon"
dynamic = ["version"]
description = ''
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
keywords = []
authors = [{ name = "Jiahao Yuan", email = "kahojyun@icloud.com" }]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = ["grpcio>=1.66.2", "protobuf>=5.27.2", "pyarrow>=17.0.0"]

[build-system]
requires = ["hatchling", "grpcio-tools==1.66.2"]
build-backend = "hatchling.build"

[project.urls]
Documentation = "https://github.com/kahojyun/fricon#readme"
Issues = "https://github.com/kahojyun/fricon/issues"
Source = "https://github.com/kahojyun/fricon"

[tool.hatch.version]
path = "src/fricon/__about__.py"

[tool.hatch.build.hooks.custom]
path = "hatch_build.py"


[tool.uv]
environments = ["python_version>='3.10'"]
dev-dependencies = [
  "grpc-stubs>=1.53.0.5",
  "grpcio-tools==1.66.2",
  "hatchling>=1.25.0",
  "ipython>=8.27.0",
  "mypy>=1.11.2",
  "pyarrow-stubs>=17.6",
  "pytest>=8.3.3",
  "ruff>=0.6.6",
  "taskipy>=1.13.0",
  "types-protobuf>=5.27.0.20240920",
]
cache-keys = [
  { file = "proto/**/*.proto" },
  { file = "hatch_build.py" },
  { file = "pyproject.toml" },
]

[tool.taskipy.tasks]
check = "mypy --install-types --non-interactive src/fricon tests"

[tool.coverage.run]
source_pkgs = ["fricon", "tests"]
branch = true
parallel = true
omit = ["src/fricon/__about__.py"]

[tool.coverage.paths]
fricon = ["src/fricon", "*/python/src/fricon"]
tests = ["tests", "*/python/tests"]

[tool.coverage.report]
exclude_lines = ["no cov", "if __name__ == .__main__.:", "if TYPE_CHECKING:"]

[tool.ruff]
extend-exclude = ["src/fricon/_proto/*"]

[tool.basedpyright]
ignore = ["src/fricon/_proto"]
